pyproject.toml
[project]
name = "llm-council"
version = "0.1.0"
description = "Your LLM Council"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "python-multipart>=0.0.20",
    "python-docx>=1.2.0",
    "pypdf>=6.4.0",
    "python-pptx>=1.0.2",
    "reportlab>=4.4.5",
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
]
